requirements.txt
GitPython==3.1.43
numpy==1.22.2
pandas==1.4.2
pathos==0.2.8
PyGithub==2.4.0
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.0.2
setuptools==58.0.4
sphinx_rtd_theme==2.0.0
sympy==1.9
